[WIP] Moving wizard to an external app
