+++ popular_proposal/views/wizard.py
@@ -19,18 +19,36 @@ from elections.models import Area
 
 from formtools.wizard.views import SessionWizardView
 
-from popular_proposal.forms import (AreaForm,
-                                    UpdateProposalForm,
+from popular_proposal.forms import (UpdateProposalForm,
                                     get_form_list,)
 
 from popular_proposal.models import (ProposalTemporaryData)
 
-from popular_proposal import send_mails_to_staff
+from popular_proposal import send_mails_to_staff, wizard_previous_form_classes
 
 wizard_form_list = get_form_list()
 
 
-class ProposalWizardBase(SessionWizardView):
+class WizardWithAreaMixin(object):
+    def determine_area(self, data):
+        is_testing = data.pop("is_testing", False)
+        if 'area' in data.keys():
+            return data['area']
+        elif hasattr(self, 'area'):
+            return self.area
+        else:
+            return Area.objects.get(id=config.DEFAULT_AREA)
+
+    def apply_extra_kwargs_from_data(self, data):
+        return {'area': self.determine_area(data)}
+
+
+class DeterminingKwargsMixin(object):
+    def apply_extra_kwargs_from_data(self, data):
+        return {}
+    
+
+class ProposalWizardBase(SessionWizardView, WizardWithAreaMixin, DeterminingKwargsMixin):
     form_list = wizard_form_list
     model = ProposalTemporaryData
     template_name = 'popular_proposal/wizard/form_step.html'
@@ -43,15 +61,6 @@ class ProposalWizardBase(SessionWizardView):
     def get_previous_forms(self):
         return []
 
-    def determine_area(self, data):
-        is_testing = data.pop("is_testing", False)
-        if 'area' in data.keys():
-            return data['area']
-        elif hasattr(self, 'area'):
-            return self.area
-        else:
-            return Area.objects.get(id=config.DEFAULT_AREA)
-
     def done(self, form_list, **kwargs):
         data = {}
         [data.update(form.cleaned_data) for form in form_list]
@@ -59,13 +68,11 @@ class ProposalWizardBase(SessionWizardView):
             'proposer': self.request.user,
             'data': data
         }
-
-        kwargs['area'] = self.determine_area(data)
+        kwargs.update(self.apply_extra_kwargs_from_data(data))
         temporary_data = self.model.objects.create(**kwargs)
         context = self.get_context_data(form=None)
-        context.update({'popular_proposal': temporary_data,
-                        'area': kwargs['area']
-                        })
+        context.update({'popular_proposal': temporary_data})
+        context.update(kwargs)
         if not settings.MODERATION_ENABLED:
             temporary_data.create_proposal()
             context['form_update'] = UpdateProposalForm(instance=temporary_data.created_proposal)
@@ -80,7 +87,7 @@ class ProposalWizardBase(SessionWizardView):
                                                                    **kwargs)
         data = self.get_all_cleaned_data()
         if data:
-            context['area'] = self.determine_area(data)
+            context.update(self.apply_extra_kwargs_from_data(data))
         context['preview_data'] = data
         return context
 
@@ -123,6 +130,8 @@ class ProposalWizard(ProposalWizardBase):
         return form_list
 
 
+wizard_initial_form_classes = wizard_previous_form_classes()
+
 class ProposalWizardFull(ProposalWizardBase):
     '''
     Esta es la clase del wizard a la que se llega por hacer
@@ -130,18 +139,16 @@ class ProposalWizardFull(ProposalWizardBase):
     Acá lo primero que te preguntamos es para qué comuna quieres crear
     la propuesta
     '''
-    form_list = [AreaForm, ] + wizard_form_list
+    form_list = wizard_initial_form_classes + wizard_form_list
 
     def get_previous_forms(self):
-        return [AreaForm, ]
+        return wizard_initial_form_classes
 
 
 class ProposalWizardFullWithoutArea(ProposalWizardBase):
     '''
     Esta es la clase del wizard a la que se llega por hacer
     /propuestas/crear
-    Acá no te preguntamos por el area por que sabemos que es la que viene por
-    defecto.
     '''
     form_list = wizard_form_list
