+++ votai_general_theme/templates/mails/popular_proposal_moderation/body.txt
@@ -11,7 +11,7 @@ Hemos leido y revisado tu propuesta ciudadana y tenemos los siguientes comentari
 {% endfor %}
 Puedes actualizar tu propuesta en el siguiente link: 
 https://{{site.domain}}{% url 'backend_citizen:temporary_data_update' pk=temporary_data.id %}
-Gracias por querer hacer de {{area}} un lugar mejor.
+Gracias.
 
 --
 El equipo de VotaInteligente.
