+++ votai_general_theme/templates/mails/popular_proposal_moderation/subject.txt
@@ -1,1 +1,1 @@
-Comentarios sobre tu propuesta ciudadana en {{area}}
+Comentarios sobre tu propuesta ciudadana: {{temporary_data.title}}
