+++ votai_general_theme/templates/popular_proposal/wizard/paso5.html
@@ -55,7 +55,9 @@
       </div>
       <div class="panel-body">
        {% bootstrap_field form.is_local_meeting  %}
+       {% comment "Sólo hasta que se arregle estaparte al final" %}
        {% bootstrap_field form.generated_at  %}
+       {% endcomment %}
       </div>
     </div>
         <section class="terms">
