+++ votainteligente/votainteligente_settings.py
@@ -25,8 +25,12 @@ AREAS_THAT_SHOULD_BE_REDIRECTING_TO_THEIR_PARENTS = ['comuna']
 SECOND_ROUND_ELECTION = None
 PRIORITY_CANDIDATES = []
 
+# Popular Proposals adaptation
 AUTHORITY_MODEL = "elections.models.Candidate"
 PROPOSALS_ADAPTER = "proposals_for_votainteligente.models.ProposalsAdapter"
 MAIL_SENDER_FUNCTION = 'votainteligente.send_mails.send_mail'
 MAIL_TO_STAFF_SENDER_FUNCTION = 'votainteligente.send_mails.send_mails_to_staff'
 DEFAULT_FITERSET_CLASS = 'proposals_for_votainteligente.filters.ProposalGeneratedAtFilter'
+WIZARD_PREVIOUS_CLASSES = ['proposals_for_votainteligente.forms.AreaForm',]
+PROPOSAL_UPDATE_FORM  = 'proposals_for_votainteligente.forms.UpdateProposalForm'
+WIZARD_FORM_MODIFIER = 'proposals_for_votainteligente.forms.wizard_forms_field_modifier'
